requirements.txt
# Schemo – Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core Framework ──────────────────────────────────────────────
Flask==3.1.0
Werkzeug==3.1.3

# ── Production Server ────────────────────────────────────────────
gunicorn==21.2.0

# ── Database ────────────────────────────────────────────────────
supabase>=2.0.0

# ── Environment Variables ────────────────────────────────────────
python-dotenv==1.0.1

# ── AI ───────────────────────────────────────────────────────────
groq>=0.9.0

# ── HTTP Requests ────────────────────────────────────────────────
requests>=2.31.0

# ── WhatsApp Notifications ────────────────────────────────────────
twilio>=8.0.0

# ── Web Scraping ──────────────────────────────────────────────
beautifulsoup4>=4.12.0
